net: modify connecitvity rematch logic to support MSIM devices
Cellular transport can have multiple agents from one or more
subscriptions. Add subscription awareness to logic while matching
requests with unspecified specifier to only match the default data
subscription agent's capabilities.

=====================================================================

multinetwork specifier match only for cellular

For Networks with non-cellular transport, we should not perform
multinetwork specifier match and return true always

Change-Id: I9c83a099a274150da95fbd4f4dbaae4e078ed4cf

=====================================================================

Mobile multisim check to be performed for Trasport type CELLULAR

Network request for WIGIG speciefier will not undergo mobile multisimm
check and will rematch with network agent with transport type WIFI

Change-Id: I7892ec6e24e93a38ae5720434772d3e8f60cbb41

=====================================================================

ConnectivityService: account for the score match.

Consider same score network as the better network if the transport
is cellular and subscription matches to current DDS.

Change-Id: Ife8fe1d005c950de64a0f518804a126a00c19ddc

=====================================================================

ConnectivityService: Extend rematch logic to prioritize network on DDS

1. Move default network to DDS always. In the multi mobile
   network scenario, give the highest priority to network on
   the current DDS irrespective of whether the network has
   low score and irrespective of the status of the non-DDS
   network.
2. Networks on non-DDS are considered unneeded and should be
   torn down.
3. Limit the changes to only multi mobile network scenarios.

Change-Id: Iadf071c44cd4cf92f5362100d608b6742a6e1fa4

=====================================================================

Clear linger if the request is for non-DDS so that it's torn down.

If there is a better network (say Wi-Fi), it replaces the cellular
network avaialble and the network starts lingering. Even if DDS
switch is triggered at this point, the cellular network which is on
non-DDS now continues to linger because unneeded returns false for
a lingering network. Fix is to clear the lingering and reap the
non-DDS networks immediately.

Change-Id: I5b1e9aeb4374d260949e0fd649700f632d08e987

CRs-Fixed: 2028848
Change-Id: Ic20d3ad34194d1bb75b25c60a3d5678dbec9a01c
Signed-off-by: Josh Fox (XlxFoXxlX) <joshfox87@gmail.com>